# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /limits:
    get:
      operationId: listLimits
      tags: [limits]
      responses:
        "200":
          description: User spending limits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Limit"
    post:
      operationId: createLimit
      tags: [limits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitInput"
      responses:
        "201":
          description: Created spending limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Limit"
  /limits/{id}:
    patch:
      operationId: updateLimit
      tags: [limits]
      parameters:
        - $ref: "#/components/parameters/LimitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitUpdate"
      responses:
        "200":
          description: Updated spending limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Limit"
        "404":
          description: Not found
    delete:
      operationId: deleteLimit
      tags: [limits]
      parameters:
        - $ref: "#/components/parameters/LimitId"
      responses:
        "204":
          description: Spending limit removed
        "404":
          description: Not found
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      responses:
        "200":
          description: User savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Savings goal details and history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDetail"
        "404":
          description: Not found
    patch:
      operationId: updateGoal
      tags: [goals]
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    delete:
      operationId: deleteGoal
      tags: [goals]
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "204":
          description: Goal removed
        "404":
          description: Not found
  /goals/{id}/movements:
    post:
      operationId: createGoalMovement
      tags: [goals]
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalMovementInput"
      responses:
        "201":
          description: Goal movement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalMovement"
        "400":
          description: Invalid movement
        "404":
          description: Not found
...
components:
  parameters:
...
    LimitId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GoalId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    LimitPeriod:
      type: string
      enum: [weekly, biweekly, monthly, bimonthly, quarterly, semiannual, annual]
    Limit:
      type: object
      required: [id, categoryId, description, amount, period, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        description: { type: ["string", "null"] }
        amount: { type: number, exclusiveMinimum: 0 }
        period:
          $ref: "#/components/schemas/LimitPeriod"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LimitInput:
      type: object
      required: [categoryId, amount, period]
      properties:
        categoryId: { type: string, format: uuid }
        description: { type: ["string", "null"] }
        amount: { type: number, exclusiveMinimum: 0 }
        period:
          $ref: "#/components/schemas/LimitPeriod"
    LimitUpdate:
      type: object
      properties:
        categoryId: { type: string, format: uuid }
        description: { type: ["string", "null"] }
        amount: { type: number, exclusiveMinimum: 0 }
        period:
          $ref: "#/components/schemas/LimitPeriod"
    Goal:
      type: object
      required: [id, title, targetAmount, savedAmount, imageData, deadline, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        savedAmount: { type: number, minimum: 0 }
        imageData: { type: ["string", "null"] }
        deadline: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GoalInput:
      type: object
      required: [title, targetAmount]
      properties:
        title: { type: string, minLength: 1 }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        imageData: { type: ["string", "null"] }
        deadline: { type: ["string", "null"], format: date }
    GoalUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        imageData: { type: ["string", "null"] }
        deadline: { type: ["string", "null"], format: date }
    GoalDetail:
      type: object
      required: [goal, history]
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
        history:
          type: array
          items:
            $ref: "#/components/schemas/GoalHistoryEntry"
    GoalHistoryEntry:
      type: object
      required: [id, type, amount, description, date, source, paymentStatus, createdAt]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [contribution, withdrawal]
        amount: { type: number, exclusiveMinimum: 0 }
        description: { type: string }
        date: { type: string, format: date }
        source:
          type: string
          enum: [transaction, manual]
        paymentStatus:
          type: string
          enum: [paid, unpaid]
        transactionType:
          type: string
          enum: [income, expense]
          nullable: true
        createdAt: { type: string, format: date-time }
    GoalMovement:
      type: object
      required: [id, goalId, type, amount, description, date, createdAt]
      properties:
        id: { type: string, format: uuid }
        goalId: { type: string, format: uuid }
        type:
          type: string
          enum: [contribution, withdrawal]
        amount: { type: number, exclusiveMinimum: 0 }
        description: { type: string }
        date: { type: string, format: date }
        createdAt: { type: string, format: date-time }
    GoalMovementInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [contribution, withdrawal]
        amount: { type: number, exclusiveMinimum: 0 }
        description: { type: string, minLength: 1 }
        date: { type: string, format: date }
...